+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Filtering stories with `StoriesParams`
 - Publishing a story with `publish()`
 - Unpublishing a story with `unpublish()`
+- Query Filters for Stories
 
 ## 0.0.1 - 2025-01-28